# api/openapi/qavia.yaml
openapi: 3.1.0
...
info:
  title: Qavia API
  version: 0.1.0
  description: |
    Internal Hyscaler platform that acts as an AI QA engineer.

    Every mutation goes through this API. The web app holds no database
    credentials and no queue access, so this is the single place authorisation is
    enforced.
  contact:
    name: Hyscaler
    email: max@hyscaler.com
servers:
  - url: /
    description: Same origin as the API process
...
security:
  - sessionCookie: []
paths:
...
  /api/v1/projects/{projectID}/test-files:
    get:
      operationId: listTestFiles
      tags: [generation]
      summary: The generated suite
      description: |
        The tree, without file content: a 300-file suite is not something to load
        whole to draw a sidebar. Paged by path, so the order a browser shows and the
        order it pages in are the same one.
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: framework
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TestFramework"
      responses:
        "200":
          description: A page of files, ordered by path.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestFilePage"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/projects/{projectID}/test-files/export:
    get:
      operationId: exportTestFiles
      tags: [generation]
      summary: Download the whole suite as a zip
      description: |
        Streamed rather than built in memory, and deterministic: the same suite
        exports to the same bytes, so one export diffs against the last.

        The archive carries a scaffold and a README, so `npm install && npm test`
        works with only environment variables supplied. Nothing in it is
        hardcoded.
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - name: framework
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TestFramework"
      responses:
        "200":
          description: A zip of the suite.
          content:
            application/zip:
              schema:
                type: string
                format: binary
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Nothing has been generated yet.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/projects/{projectID}/test-files/postman:
    post:
      operationId: exportPostmanCollection
      tags: [generation]
      summary: Build a Postman collection from approved cases
      description: |
        Deterministic code rather than an agent: a collection is JSON with a known
        shape, and the approved cases already say what to call and what to assert.

        Base URL and auth are collection variables, not literals, so one export
        runs against any environment.
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "201":
          description: The collection, stored as a file in the suite.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestFile"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: No approved test cases to export.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: no_approved_test_cases
                message: No test cases are approved yet. Approve the ones you want implemented, then generate code.
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/test-files/{testFileID}:
    get:
      operationId: getTestFile
      tags: [generation]
      summary: One file, with its content
      description: |
        Content is plain text. Highlighting is a frontend concern, and returning
        pre-rendered HTML would make the API responsible for a rendering decision
        it cannot see.
      parameters:
        - $ref: "#/components/parameters/TestFileID"
      responses:
        "200":
          description: The file and the cases it covers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestFileContent"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/test-files/{testFileID}/download:
    get:
      operationId: downloadTestFile
      tags: [generation]
      summary: Download one file
      parameters:
        - $ref: "#/components/parameters/TestFileID"
      responses:
        "200":
          description: The file.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/test-cases/{testCaseID}/files:
    get:
      operationId: listFilesForTestCase
      tags: [generation]
      summary: Where a test case is implemented
      description: |
        The traceability question in the other direction (F-6.11): given a case,
        which generated files implement it.
      parameters:
        - $ref: "#/components/parameters/TestCaseID"
      responses:
        "200":
          description: The files covering this case.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestFileList"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: qavia_session
      description: |
        Server-side session. Sessions live in Postgres rather than in a JWT, so
        revocation is immediate: an offboarded employee's access ends when the row
        is deleted, with no token TTL to wait out.
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      description: Page size. Every list has a hard maximum; there is no unbounded list.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor from a previous response's nextCursor. Do not construct one.
      schema:
        type: string
        maxLength: 512
...
    ProjectID:
      name: projectID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    TestCaseID:
      name: testCaseID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TestFileID:
      name: testFileID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    Error:
      type: object
      description: |
        The single error shape. Produced only by the mapper in
        internal/platform/httpx, so every failure looks the same.
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            Stable and machine readable. Branch on this, never on message. Every
            value is catalogued in docs/error-codes.md.
          example: target_host_not_allowed
        message:
          type: string
          description: Safe to show a user, and says what to do next.
          example: Host "api.example.test" is not on this project's allowlist. Add it in Settings, Project, Targets.
        details:
          type: object
          additionalProperties: true
          description: Field-level context for inline rendering. Never carries a secret.
          example:
            host: api.example.test
        incidentId:
          type: string
          description: |
            Present only on a 500. The handle between what the user saw and the
            full wrapped cause in the log, which is never returned.
          example: 9f2c1a7e
...
    NextCursor:
      type: [string, "null"]
      description: Pass as cursor to fetch the next page. Null or absent means the last page.
      example: eyJjIjoiMjAyNi0wOC0xMFQwOToxNDoyMloifQ
...
    TestFramework:
      type: string
      description: |
        What a file is written for. Supertest and Postman are delivered in this
        phase; the rest are declared so the API and the UI agree about what exists,
        and asking for one returns a reason rather than nothing.
      enum: [supertest, postman, jest, vitest, playwright, cypress, k6, pytest]
    TestFile:
      type: object
      description: One generated file, without its content.
      required: [id, projectId, framework, path, testCaseIds, sizeBytes, generatedAt]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        framework:
          $ref: "#/components/schemas/TestFramework"
        path:
          type: string
          description: Repository-relative and forward-slashed. Validated before storage.
        testCaseIds:
          type: array
          description: The cases this file implements, which is what makes traceability work.
          items:
            type: string
            format: uuid
        generatedBy:
          type: string
        validatedAt:
          type: [string, "null"]
          format: date-time
          description: |
            Null means static validation has not run yet, which is not the same as
            having failed.
        validationNote:
          type: string
        sizeBytes:
          type: integer
        generatedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TestFileContent:
      allOf:
        - $ref: "#/components/schemas/TestFile"
        - type: object
          required: [content]
          properties:
            content:
              type: string
              description: Plain text. Highlighting is a frontend concern.
...
    TestFilePage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TestFile"
        nextCursor:
          $ref: "#/components/schemas/NextCursor"
        summary:
          type: array
          description: File and byte counts per framework, so a browser can show totals without paging.
          items:
            $ref: "#/components/schemas/TestFileSummary"
    TestFileSummary:
      type: object
      required: [framework, files, bytes]
      properties:
        framework:
          $ref: "#/components/schemas/TestFramework"
        files:
          type: integer
          format: int64
        bytes:
          type: integer
          format: int64
    TestFileList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TestFile"
  responses:
...
    Unauthorized:
      description: No valid session.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: unauthenticated
            message: Sign in to continue.
    Forbidden:
      description: Authenticated, but not permitted.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: role_required
            message: This action requires the Admin role.
    NotFound:
      description: |
        Not found, or not visible to this caller. The two are deliberately
        indistinguishable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: project_not_found
            message: Project not found.
...
    InternalError:
      description: |
        An unexpected failure. Carries an incident ID and nothing else; the cause
        is in the structured log against that ID.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: internal_error
            message: Something went wrong on our side.
            incidentId: 9f2c1a7e